Refactor(scripts): Split the two oversized scripts
audit_templates.py (513 lines) and compare_github_languages.py (644
lines) both sat past the 400-line reviewability limit the aislop
code-quality gate applies.

Keep both entry points at their documented paths, so 'python
scripts/audit_templates.py' still works, and move their helpers into
sibling modules: template_audit_analysis, template_audit_fixtures and
template_audit_reporting for the audit; github_language_sources and
github_language_comparison for the comparison. Each entry point puts
its own directory on sys.path before importing the siblings, so the
scripts run from any working directory as before.
